backend/requirements.txt
fastapi
uvicorn
supabase
pinecone
openai
pydantic
python-dotenv
python-multipart
langchain
langchain-openai
langchain-community
langgraph
PyPDF2
python-jose[cryptography]
passlib[bcrypt]
composio-langchain
cohere
pytest
pytest-asyncio
httpx
yt-dlp
feedparser
twikit
openai-whisper
youtube-transcript-api
